FAC-983: Refactor - Move `Service`, `ServiceTest` and `useCase` to correct package
